Revert "Changing key type of vertex_colors from uint32_t to int64_t"
This reverts commit a9854c65019a19dc2ac98d7245bc3093cae0e87f.

+++ scene/2d/polygon_2d.h
@@ -35,6 +35,8 @@
 #include "core/variant/typed_dictionary.h"
 #include "scene/2d/node_2d.h"
 
+#include <cstdint>
+
 class NavigationPolygon;
 class NavigationMeshSourceGeometryData2D;
 
@@ -43,7 +45,7 @@ class Polygon2D : public Node2D {
 
 	Vector<Vector2> polygon;
 	Vector<Vector2> uv;
-	TypedDictionary<int64_t, Color> vertex_colors;
+	TypedDictionary<uint32_t, Color> vertex_colors;
 	Array polygons;
 	int internal_vertices = 0;
 
@@ -125,8 +127,8 @@ public:
 	void set_color(const Color &p_color);
 	Color get_color() const;
 
-	void set_vertex_colors(const TypedDictionary<int64_t, Color> &p_colors);
-	TypedDictionary<int64_t, Color> get_vertex_colors() const;
+	void set_vertex_colors(const TypedDictionary<uint32_t, Color> &p_colors);
+	TypedDictionary<uint32_t, Color> get_vertex_colors() const;
 
 	void set_texture(const Ref<Texture2D> &p_texture);
 	Ref<Texture2D> get_texture() const;
